+++ lib/Settings/PersonalSection.php
@@ -1,0 +1,35 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OCA\Attendance\Settings;
+
+use OCP\IL10N;
+use OCP\IURLGenerator;
+use OCP\Settings\IIconSection;
+
+class PersonalSection implements IIconSection {
+	private IL10N $l;
+	private IURLGenerator $urlGenerator;
+
+	public function __construct(IL10N $l, IURLGenerator $urlGenerator) {
+		$this->l = $l;
+		$this->urlGenerator = $urlGenerator;
+	}
+
+	public function getID(): string {
+		return 'attendance';
+	}
+
+	public function getName(): string {
+		return $this->l->t('Attendance');
+	}
+
+	public function getPriority(): int {
+		return 75;
+	}
+
+	public function getIcon(): string {
+		return $this->urlGenerator->imagePath('attendance', 'app-dark.svg');
+	}
+}